Changes Bluespace Sprite to the Universally Liked One
Interestingly Nexis *accidentally* changed his PR to put the sprite that everyone supported his in PR behind a define, and put the flat texture back in by accident (I assume because he didn't say anything). This rights that mistake and undoes the define so the new texture in his PR is the bluespace texture.